// src/Enums/MonsterTypeEnum.ts
export enum MonsterTypeEnum {
  Aerocephal = 'aerocephal',
  Arcana_Drake = 'arcana_drake',
  Aurum_Drakueli = 'aurum_drakueli',
  Bat = 'bat',
  Daemarbora = 'daemarbora',
  Deceleon = 'deceleon',
  DemonicEssence = 'demonic_essence',
  DuneCrawler = 'dune_crawler',
  GreenSlime = 'green_slime',
  Nagaruda = 'nagaruda',
  Rat = 'rat',
  Scorpion = 'scorpion',
  Skeleton = 'skeleton',
  Snake = 'snake',
  Spider = 'spider',
  StygianLizard = 'stygian_lizard'
}